rendered different views for large and small screens
