+++ client/src/App.js
@@ -1,94 +1,39 @@
 import React, { Component } from "react";
-import { findDOMNode } from "react-dom"
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import { connect } from "react-redux";
+import { BrowserRouter as Router } from "react-router-dom";
 import { CssBaseline } from "@material-ui/core";
 import { MuiThemeProvider } from "@material-ui/core/styles";
-import SplitPane from "react-split-pane";
-import { Scrollbars } from "react-custom-scrollbars";
-import injectSheet from "react-jss"
+import $ from "jquery";
 
-import theme, { secondary } from "./theme";
-import { resizePane } from "./actions/splitPaneActions";
-import { startWaiting, sendUpdateSignal } from "./actions/projectListActions";
+import theme from "./theme";
 import { NavBar, ResumeFloatingButton } from "./components/Menus";
-import AboutMe from "./components/AboutMe";
-import Projects from "./components/Projects";
-import ProjectDetail from "./components/ProjectDetail";
-import Resume from "./components/Resume";
-
-const styles = {
-  resizer: {
-    cursor: "col-resize",
-    width: 7,
-    backgroundColor: secondary,
-  },
-}
-
-
-const mapStateToProps = store => ({ splitPane: store.splitPane });
+import LargeScreen from "./components/LargeScreen";
+import SmallScreen from "./components/SmallScreen";
 
 class App extends Component {
   state = {
-    pane1Style: { transition: "width 0.3s ease" },
-  }
-
-  constructor(props) {
-    super(props);
-    this.splitPane = React.createRef();
+    screen: "large"
   }
 
   componentDidMount() {
-    const node = findDOMNode(this.splitPane.current.pane1);
-    node.addEventListener("transitionend", event => {
-      if (event.propertyName === "width") {
-        this.props.dispatch(sendUpdateSignal());
-      }
-    });
-  }
-
-  handlePaneResize = width => this.props.dispatch(resizePane(width));
-
-  handleDragStarted = () => {
-    this.setState({ pane1Style: undefined, resizerStyle: { opacity: 0.8 } });
-    this.props.dispatch(startWaiting());
-  }
-
-  handleDragFinished = () => {
-    this.setState({ pane1Style: { transition: "width 0.3s linear" }, resizerStyle: undefined });
-    this.props.dispatch(sendUpdateSignal());
+    this.setState({
+      screen: $(window).width() > 900 ? "large" : "small"
+    })
   }
 
   render() {
-    const { splitPane, classes } = this.props;
+    const { screen } = this.state;
     return (
       <MuiThemeProvider theme={theme}>
         <Router>
           <div>
             <CssBaseline />
             <NavBar />
-            <ResumeFloatingButton />
-            <SplitPane
-              ref={this.splitPane}
-              split="vertical"
-              onChange={this.handlePaneResize}
-              onDragStarted={this.handleDragStarted}
-              onDragFinished={this.handleDragFinished}
-              minSize={300}
-              size={splitPane.open ? splitPane.size : "100%"}
-              pane1Style={this.state.pane1Style}
-              resizerStyle={this.state.resizerStyle}
-              resizerClassName={classes.resizer}
-            >
-              <Switch>
-                <Route exact path="/" component={AboutMe}/>
-                <Route exact path="/projects" component={Projects}/>
-                <Route exact path="/projects/:slug" component={ProjectDetail}/>
-              </Switch>
-              <Scrollbars>
-                <Resume />
-              </Scrollbars>
-            </SplitPane>
+            <ResumeFloatingButton screen={screen} />
+            { screen === "large" ? (
+              <LargeScreen />
+            ) : (
+              <SmallScreen />
+            )}
           </div>
         </Router>
       </MuiThemeProvider>
@@ -96,4 +41,4 @@ class App extends Component {
   }
 }
 
-export default connect(mapStateToProps)(injectSheet(styles)(App));
+export default App;
